+++ src/gas/routes.ts
@@ -1,11 +1,11 @@
 import Utils from "./shared/Utils";
-import Get from "./controllers/clubs/Get";
-import Regist from "./controllers/club/Regist";
-import Approve from "./controllers/club/Approve";
-import Join from "./controllers/club/Join";
-
 import ParameterInterface from "./types/ParameterInterface";
 import ResponseInterface from "./types/ResponseInterface";
+
+import GetController from "./controllers/Clubs/GetController";
+import RegistController from "./controllers/Club/RegistController";
+import ApproveController from "./controllers/Club/ApproveController";
+import JoinController from "./controllers/Club/JoinController";
 
 interface DoGetParams extends GoogleAppsScript.Events.DoGet {
   parameter: ParameterInterface;
@@ -18,12 +18,12 @@ interface DoPostParams extends GoogleAppsScript.Events.DoPost {
 const doGet = (e: DoGetParams) => {
   const { action } = e.parameter;
 
-  const get = new Get();
+  const get = new GetController();
   const utils = new Utils();
 
   switch (action) {
     case "get": {
-      return utils.createOutput(get.read());
+      return utils.createOutput(get.show());
     }
     default: {
       return utils.createOutput(utils.makeError({ status: 404, message: "404 Not Found" }));
@@ -36,18 +36,20 @@ const doPost = (e: DoPostParams) => {
   // FIXME: これで取れる？
   const params = JSON.parse(e.postData.getDataAsString());
 
-  const regist = new Regist();
-  const approve = new Approve();
-  const join = new Join();
+  const regist = new RegistController();
+  const approve = new ApproveController();
+  const join = new JoinController();
   const utils = new Utils();
 
   switch (action) {
     case "regist": {
       return utils.createOutput(regist.create(params));
     }
+    // WIP
     case "approve": {
       return utils.createOutput(approve.create(params));
     }
+    // WIP
     case "join": {
       return utils.createOutput(join.update(params));
     }
